Mod hal_led Config
Modify the Port of LED Initialization on HAL_BOARD_INIT() what is macro
define on hal_board_cfg.h.
Modify the LED Port Configuration on the hal_board_cfg.h.
Modify the control macro define of LED on hal_borad_cfg.h

Delete the configuration of LED which is not be used on hal_borad_cfg.h.
Delete the Port initialization of LED which is not be used on
HAL_BOARD_INIT()
Delete the marco compiler which will never be used on hal_borad_cfg.h