Mod the GenericApp
Modify the GenericApp_HandleKeys() let the oled can show the panid of the ECG
connect when link key pressed.
Delete the feature of ECG device send message every 5 seconds when the
network changed.
This version is ECG-beta0.1 which is can connecting to
GW-withoutSTM32-beta0.1.
The feature of this version just show the panid of the ECG device
connection.